Created framework for different windows
Created a windowmanager class which has the frame as an attribute and creates the other windows as objects of the other classes (MenuWindow and SettingsWindow). Has the capability to switch between the two at the moment on command but doesnt yet. Next step is to take the buttonListener class out of the menu window so that it can be used by all classes - this will allow the buttons to be used to change between windows (cards in cardlayout).