Stats are now unique to users
The login page calls a function to change the menu's variable of which user is currently logged in. The stats are read off that user's file (with no encryption) and essentially calculates the best and average score based off of a list of scores delimited by commas. An issue is a numberformat exception when ",," is present. Also it is recalculated every time a user clicks a difficulty, but this is inefficient, so this could be done every 5 turns or so once the actual game is implemented.