Game window now generates multiple questions
it does so until the time runs out, which is set as 10 seconds at the moment. Your score is added up and passed as an argument to the menu class when it goes back to it. This is then stored temporarily in an array for display. Added a refresh button that does not work, formatting on the infoPanel needs to be sorted and score values from the game need to be written in to the appropriate files.