Game now writes scores into files
These scores are now read correctly by the menu to be written in the info panel. Game is now fully functional, but there are still things to be done:
Input box should be cleared after question is answered in game class.
Issue once logged in the data from the easy line is read but not the rest.
The alt text before a difficulty is selected "select difficulty" does not appear. Additionally, there is nothing in the settings page and also the game window needs to be graphically designed.