Improved operator and operand selection
FIxed issue where operator and second operand would not display, and made a list of lists for operators and a list of list of ranges of values to use . These are then selected based off of the difficulty value and the random integer that is generated. Then the operand is generated as a random value between the range selected.